remove pdf generator singlton, do not want to have too many instances for memory
